battalion consumption.

+++ empery_center/src/data/map_object_type.cpp
@@ -98,7 +98,7 @@ namespace {
 			elem.maintenance_cost.reserve(object.size());
 			for(auto it = object.begin(); it != object.end(); ++it){
 				const auto resource_id = boost::lexical_cast<ResourceId>(it->first);
-				const auto resource_amount = static_cast<std::uint64_t>(it->second.get<double>());
+				const auto resource_amount = it->second.get<double>();
 				if(!elem.maintenance_cost.emplace(resource_id, resource_amount).second){
 					LOG_EMPERY_CENTER_ERROR("Duplicate maintenance resource cost: resource_id = ", resource_id);
 					DEBUG_THROW(Exception, sslit("Duplicate maintenance resource cost"));
